+++ server/db/update_v1.1.6_v1.1.7.sql
@@ -1,0 +1,2 @@
+SET @id_style = (SELECT `id` FROM `styles` WHERE `name` = 'userProgress');
+UPDATE `styles` SET `id_type` = '0000000002' WHERE `styles`.`id` = @id_style;